requirements.txt
fastapi==0.78.0
keras==2.8.0
numpy==1.22.3
pandas==1.4.1
Pillow==9.2.0
tensorflow==2.8.0